requirements/requirements.txt
alembic==1.7.6
amqp==5.0.9
billiard==3.6.4.0
blinker==1.4
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
ForgeryPy3==0.3.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
prompt-toolkit==3.0.27
psycopg2==2.9.3
pyparsing==3.0.7
pytz==2021.3
redis==4.1.2
six==1.16.0
SQLAlchemy==1.4.31
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.1
zipp==3.7.0
